workers/document/extractor/requirements.txt
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
fpdf2>=2.8.0
markdown>=3.6.0
